add extremely primitive 'connect' functionality
WOW today was amazing. I'm also really enjoying actually writing my
commit messages in vim. It practically *incentivizes* absurdly long
messages -- which works for me since I'm using them as a pseudo-journal.

OK so today was great. We got a TCP connection. But, hilariously, there
seems not to be a canonical (or maybe even *effective*?) way to test
whether a TCP connection is still alive in golang. Currently the port is
hardcoded, which is nonideal. I'd really lilke it if the client
tries first to get that info from the SRV record for that domain since
afaik that's the preferred way as per the XMPP spec. But yeah. THIS IS
REAL HEADWAY AND I'M SO STOKED!

Last week was rough -- minimal progress on this project (or so it felt),
and I spread myself too thin with non-code obligations. However, over
the weekend I went to an *amazing* techno show with an RC friend and had
an awesome time. I feel rejuvenated.