remove replace libclaster with go-xmpp
LOOK AT IT. IT'S A FUNCTIONING CONNECTION TO A SERVER. This is great.
This commit scraps my xmpp library and replaces with go-xmpp from mattn.
So far the client is able to log in to a server. Currently it only
supports connections to servers that have STARTTLS support. I think this
is fine.

Today has been great so far. I've got this commit, which is a very happy
commit, and I'm working downstairs at one of the pairing stations. This
means that A) I've had a giant external monitor. Freakin' lovely. And B)
Lots of people have come up to ask what I'm working on and to explain to
them the specific thing I'm doing at the moment that they walk up. Which
has been lovely! It feels great to have finally made enough headway on
my project that I can talk about it with people.

I haven't paired with anyone on claster yet, but I did pair with another
recurser on something that she was working on. The application is a
flask app and is in production now. I mostly watched and learned about
python/flask, but there were one or two bugs that I may have had a hand
in triaging :)  Also, she introduced me to two VSCode extensions which
are so so amazing (indent-rainbow and rainbow brackets) -- this
triggered me to actually commit some energy to themeing VSCode a bit and
so I did and now I'm really happy with it!

Also, this person introduced me to Ctrl+r in bash. WHERE HAS THIS BEEN
ALL MY LIFE.

This week started lowish and appears to be ending high. WOOO